added dev dependency handling and workflow and updates to package ranges
