+++ .gitignore
@@ -207,6 +207,7 @@ WARP.md
 **/tmpclaude*
 # Dependency-bound validation reports
 python/scripts/dependency-*-results.json
+python/scripts/dependencies/dependency-*-results.json
 
 # Azurite storage emulator files
 */__azurite_db_blob__.json*
