+++ python/packages/devui/pyproject.toml
@@ -30,8 +30,15 @@ dependencies = [
 ]
 
 [project.optional-dependencies]
-dev = ["pytest>=7.0.0", "watchdog>=3.0.0", "agent-framework-orchestrations"]
-all = ["pytest>=7.0.0", "watchdog>=3.0.0"]
+dev = [
+    "pytest==9.0.2",
+    "watchdog==6.0.0",
+    "agent-framework-orchestrations==1.0.0b260304",
+]
+all = [
+    "pytest==9.0.2",
+    "watchdog==6.0.0",
+]
 
 [project.scripts]
 devui = "agent_framework_devui:main"
