+++ python/packages/durabletask/AGENTS.md
@@ -29,18 +29,16 @@ Durable execution support for long-running agent workflows using Azure Durable F
 ## Usage
 
 ```python
-from durabletask.client import TaskHubGrpcClient
-from durabletask.worker import TaskHubGrpcWorker
-from agent_framework import ChatAgent
+from agent_framework import Agent
 from agent_framework.azure import AzureOpenAIChatClient
 from agent_framework_durabletask import DurableAIAgentClient, DurableAIAgentWorker
+from durabletask.client import TaskHubGrpcClient
+from durabletask.worker import TaskHubGrpcWorker
 
 # Client side
 dt_client = TaskHubGrpcClient(host_address="localhost:4001")
 agent_client = DurableAIAgentClient(dt_client)
-agent = agent_client.get_agent("assistant")
-response = agent.run("Hello, how are you?")
-print(response.text)
+durable_agent = agent_client.get_agent("assistant")
 
 # Worker side
 dt_worker = TaskHubGrpcWorker(host_address="localhost:4001")
@@ -48,10 +46,8 @@ agent_worker = DurableAIAgentWorker(dt_worker)
 
 # Create a chat client for the agent
 chat_client = AzureOpenAIChatClient()
-my_agent = ChatAgent(chat_client=chat_client, name="assistant")
+my_agent = Agent(client=chat_client, name="assistant")
 agent_worker.add_agent(my_agent)
-
-dt_worker.start()
 ```
 
 ## Import Path
