+++ python/packages/durabletask/pyproject.toml
@@ -30,7 +30,7 @@ dependencies = [
 
 [dependency-groups]
 dev = [
-    "types-python-dateutil>=2.9.0",
+    "types-python-dateutil==2.9.0.20260305",
 ]
 
 [tool.uv]
