+++ python/packages/mem0/AGENTS.md
@@ -4,23 +4,25 @@ Integration with Mem0 for agent memory management.
 
 ## Main Classes
 
-- **`Mem0Provider`** - Context provider that integrates Mem0 memory into agents
+- **`Mem0ContextProvider`** - Context provider that integrates Mem0 memory into agents
 
 ## Usage
 
 ```python
-from agent_framework.mem0 import Mem0Provider
+from agent_framework.mem0 import Mem0ContextProvider
 
-provider = Mem0Provider(api_key="your-key")
-agent = Agent(..., context_provider=provider)
+provider = Mem0ContextProvider(
+    api_key="your-key",
+    user_id="user-id",
+)
 ```
 
 ## Import Path
 
 ```python
-from agent_framework.mem0 import Mem0Provider
+from agent_framework.mem0 import Mem0ContextProvider
 # or directly:
-from agent_framework_mem0 import Mem0Provider
+from agent_framework_mem0 import Mem0ContextProvider
 ```
 
 ## Notes
