+++ python/scripts/__init__.py
@@ -1,0 +1,5 @@
+# Copyright (c) Microsoft. All rights reserved.
+
+from __future__ import annotations
+
+"""Shared Python workspace scripts."""
