+++ python/scripts/dependencies/__init__.py
@@ -1,0 +1,5 @@
+# Copyright (c) Microsoft. All rights reserved.
+
+from __future__ import annotations
+
+"""Dependency maintenance and validation scripts for the Python workspace."""
